:sparkles: initial worker reconcile loop
