+++ config/samples/infrastructure_v1alpha1_worker.yaml
@@ -4,3 +4,4 @@ metadata:
   name: worker-sample
 spec:
   capacity: 2
+  location: southcentralus
