+++ go.mod
@@ -11,8 +11,7 @@ require (
 	github.com/onsi/ginkgo v1.12.0
 	github.com/onsi/gomega v1.9.0
 	github.com/prometheus/common v0.9.1
-	github.com/spf13/pflag v1.0.5
-	golang.org/x/sys v0.0.0-20200302150141-5c8b2ff67527 // indirect
+	go.uber.org/zap v1.10.0
 	k8s.io/api v0.17.4
 	k8s.io/apimachinery v0.17.4
 	k8s.io/cli-runtime v0.17.4
